requirements.txt
asgiref==3.8.1
certifi==2025.6.15
charset-normalizer==3.4.2
django==5.2.3
django-anymail==13.0
django-cors-headers==4.7.0
django-unfold==0.60.0
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
idna==3.10
markdown==3.8.2
pillow==11.3.0
pyjwt==2.9.0
pyotp==2.9.0
python-decouple==3.8
requests==2.32.4
sqlparse==0.5.3
urllib3==2.5.0
